rpc: add GetProtocolVersion alpha endpoint
