+++ crates/sui-rpc-api/proto/sui.node.v2alpha.proto
@@ -51,6 +51,8 @@ service NodeService {
 
   // List the dynamic fields for provided parent.
   rpc ListDynamicFields(ListDynamicFieldsRequest) returns (ListDynamicFieldsResponse);
+
+  rpc GetProtocolConfig(GetProtocolConfigRequest) returns (GetProtocolConfigResponse);
 }
 
 // Request message for `NodeService.GetCoinInfo`.
@@ -160,4 +162,17 @@ message DynamicField {
   // The presence or absence of this field can be used to determine if a child
   // is a dynamic field or a dynamic child object
   optional sui.types.ObjectId dynamic_object_id = 5;
+}
+
+message GetProtocolConfigRequest {
+  optional uint64 version = 1;
+}
+
+message GetProtocolConfigResponse {
+  optional uint64 protocol_version = 1;
+  map<string, bool> feature_flags = 2;
+  map<string, string> attributes = 3;
+
+  optional uint64 max_suppported_protocol_version = 4;
+  optional uint64 min_suppported_protocol_version = 5;
 }
